Refactor atlas: extract UI primitives and split large components
Adds <FloatingPanel>, <MapControlButton>, <MenuItemRow>, style tokens,
and a shared formatters module; splits inspector.tsx (929->185),
layer-toggle.tsx (604->294), and airspace-layer.tsx (779->434) into
focused modules. No behavior change; 264 tests pass.
